requirements.txt
python==3.9.7
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.22.3
andas==1.4.1
pickleshare==0.7.5
Pillow==9.0.1
scikit-learn==1.0.2
seaborn==0.11.2
treamlit==1.7.0